Версия с разделением кода
