+++ figure.mjs
@@ -1,0 +1,54 @@
+ const figure1 = [
+    [0, 0, 0, 0],
+    [0, 1, 1, 0],
+    [0, 1, 1, 0],
+    [0, 0, 0, 0],
+];
+ const figure2 = [
+    [0, 0, 0, 0],
+    [0, 1, 1, 0],
+    [0, 1, 0, 0],
+    [0, 1, 0, 0],
+];
+ const figure3 = [
+    [0, 1, 0, 0],
+    [0, 1, 0, 0],
+    [0, 1, 0, 0],
+    [0, 1, 0, 0],
+];
+ const figure4 = [
+    [0, 0, 0, 0],
+    [1, 1, 1, 0],
+    [0, 1, 0, 0],
+    [0, 0, 0, 0],
+];
+ const figure5 = [
+    [0, 0, 0, 0],
+    [0, 1, 0, 0],
+    [0, 1, 1, 0],
+    [0, 0, 1, 0],
+];
+ const figure6 = [
+    [0, 0, 0, 0],
+    [1, 1, 0, 0],
+    [0, 1, 1, 0],
+    [0, 0, 0, 0],
+];
+ const figure7 = [
+    [0, 0, 0, 0],
+    [1, 1, 1, 0],
+    [1, 0, 0, 0],
+    [0, 0, 0, 0],
+];
+ const test = 44;
+ 
+export {
+    figure1,
+    figure2,
+    figure3,
+    figure4,
+    figure5,
+    figure6,
+    figure7,
+    test
+};
