--- game.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Tetris</title>
-    <link rel="stylesheet" href="game.css">
-  </head>
- <center>
-	<body>
-		<table cellspacing="0" class="layout">
-			<tr> 
-				<td class="leftcol">				
-					<h1 >Tetris</h1>
-					<canvas id="canvasid" height="600" width="300"></canvas> 	
-				</td>
-				<td class="rightcol">	
-						</br></br></br></br></br></br>
-					<button id = "pause" >Пауза</button> <button id = "restart" >Рестарт Tetris</button> 
-					</br></br>
-					<button id = "plusSpeed" >+ к скорости игры </button> <button id = "minusSpeed" >- к скорости игры</button>   
-					<p>Счет - <span id="count_row">0</span> <p>
-					<p> Скорость игры: &nbsp <output id = "outSpeed" value = "1">1</output></p>					
-					<canvas id = "nextFigure" height="120" width="120"></canvas>
-				</td>
-			</tr>
-		</table>
-	</body>
- </center>
-  
-   <script src="game.js"></script>
-   <button class="play">Play Music</button>
-   <input type="range" id="volumeSlider" min="0" max="1" step="0.1" value="0.5">
-   <audio id="backgroundSound" src="./audio/game.ogg" loop></audio>
-   <button class="pause">Shut Down Music</button>
-   <script src="./audio.js"></script>
-</html>
