+++ package.json
@@ -1,10 +1,12 @@
 {
   "scripts": {
+    "type": "module",
     "lint": "eslint . --ext .js",
     "lint-fix": "eslint . --ext .js --fix"
   },
   "devDependencies": {
     "eslint": "^8.40.0",
-    "eslint-config-google": "^0.14.0"
+    "eslint-config-google": "^0.14.0",
+    "type": "module"
   }
 }
